+++ src/main/VPSBuyer.py
@@ -1,0 +1,58 @@
+from random import randint
+from selenium import webdriver
+from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
+from selenium.webdriver.support.select import Select
+from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
+from selenium.common.exceptions import TimeoutException
+
+from main.BogusFormBuilder import BogusFormBuilder
+
+
+class VPSBuyer(object):
+    '''
+    This is the standard class to buy a VPS host. By itself, it does nothing; this class is supposed to be extended by other classes, each for a specific VPS Provider
+    '''
+    def __init__(self):
+        self.driver = webdriver.Remote(
+            command_executor='http://127.0.0.1:4444/wd/hub',
+            desired_capabilities=DesiredCapabilities.FIREFOX)
+        self.generator = BogusFormBuilder()
+        
+        self.email = self.generator.getEmail()
+        #password = generator.getPassword()
+        self.password = self.generator.getRAString(32)
+        
+        self.SSHUsername = ""
+        self.SSHPassword = ""
+        pass
+    
+    def getFormValue(self, name):
+       "function_docstring"
+       return "To be implemented: " + name
+    
+    
+    '''
+    Automatically fills ina form element by executing a piece of javascript that sets the value attribute of the form element
+    '''
+    def fillInElement(self, fieldname, value):
+        #driver.find_element_by_css_selector("input[name='" + fieldname + "']").send_keys(value)
+        
+        # ^ send_keys has some issues, using javascript to set an attribute instead:
+        self.driver.execute_script('arguments[0].setAttribute("value", "' + value + '")', self.driver.find_element_by_css_selector("input[name='" + fieldname + "']"))
+        
+    '''
+    Chooses one of the elements in a select list randomly, except for the first element
+    '''
+    def clickRandomSelectElement(self, fieldId):
+        el = self.driver.find_element_by_id(fieldId)
+        options = el.find_elements_by_tag_name('option')
+        num = randint(1, len(options) - 1)
+        option = options[num]
+        option.click()
+        
+    def getSSHUsername(self):
+        return self.SSHUsername
+    
+    def getSSHPassword(self):
+        return self.SSHPassword
